+++ src/mapfile_parser/__init__.py
@@ -5,7 +5,7 @@
 
 from __future__ import annotations
 
-__version_info__ = (2, 8, 0)
+__version_info__ = (2, 9, 0)
 __version__ = ".".join(map(str, __version_info__)) + "-dev0"
 __author__ = "Decompollaborate"